requirements.txt
google-api-python-client
gspread
oauth2client
ShopifyAPI
